refactor: 컬러박스들 가져오는 로직을 BrushColorBox로 이동
