+++ app/src/main/java/woowacourse/paint/customview/paint/BrushColor.kt
@@ -2,7 +2,6 @@ package woowacourse.paint.customview.paint
 
 import androidx.annotation.ColorRes
 import woowacourse.paint.R
-import woowacourse.paint.main.model.BrushColorBox
 
 enum class BrushColor(@ColorRes val colorRes: Int) {
     RED(R.color.red),
@@ -11,15 +10,4 @@ enum class BrushColor(@ColorRes val colorRes: Int) {
     GREEN(R.color.green),
     BLUE(R.color.blue),
     ;
-
-    companion object {
-        fun getColorBoxes(selectedColor: BrushColor): List<BrushColorBox> =
-            BrushColor.values().map { brushColor ->
-                if (selectedColor == brushColor) {
-                    BrushColorBox(brushColor, true)
-                } else {
-                    BrushColorBox(brushColor, false)
-                }
-            }
-    }
 }
