+++ app/src/main/java/woowacourse/paint/main/model/BrushColorBox.kt
@@ -5,4 +5,15 @@ import woowacourse.paint.customview.paint.BrushColor
 data class BrushColorBox(
     val brushColor: BrushColor,
     val isSelected: Boolean = false,
-)
+) {
+    companion object {
+        fun getColorBoxes(selectedColor: BrushColor): List<BrushColorBox> =
+            BrushColor.values().map { brushColor ->
+                if (selectedColor == brushColor) {
+                    BrushColorBox(brushColor, true)
+                } else {
+                    BrushColorBox(brushColor, false)
+                }
+            }
+    }
+}
